rename: common.ts -> auth.ts로 파일명 변경
